fix(*): resolved all build errors
