+++ src/commands/General/User Settings/userconf.ts
@@ -1,11 +1,10 @@
-import { Argument, Command, CommandStore, GatewayStorage, Schema, SchemaEntry, SchemaFolder, SettingsFolder } from 'klasa';
-import { toTitleCase, codeBlock } from '@klasa/utils';
+import { Argument, Command, CommandStore, Settings, SchemaEntry, Gateway } from 'klasa';
+import { toTitleCase } from '@klasa/utils';
+import { codeblock } from 'discord-md-tags';
 
 import type { Message, Guild } from '@klasa/core';
 
 export default class extends Command {
-
-	private readonly configurableSchemaKeys = new Map<string, Schema | SchemaEntry>();
 
 	public constructor(store: CommandStore, directory: string, files: readonly string[]) {
 		super(store, directory, files, {
@@ -28,25 +27,24 @@ export default class extends Command {
 			});
 	}
 
+	private get gateway(): Gateway {
+		return this.client.gateways.get('users') as Gateway;
+	}
+
 	public show(message: Message, [key]: [string]): Promise<Message[]> {
-		const schemaOrEntry = this.configurableSchemaKeys.get(key);
-		if (typeof schemaOrEntry === 'undefined') throw message.language.get('COMMAND_CONF_GET_NOEXT', key);
+		if (!key) return message.sendLocale('COMMAND_CONF_SERVER', [key, codeblock('asciidoc')`${this.displayFolder(message.author.settings)}`]);
 
-		const value = key ? message.author.settings.get(key) : message.author.settings;
-		if (SchemaEntry.is(schemaOrEntry)) {
-			return message.replyLocale('COMMAND_CONF_GET', [key, this.displayEntry(schemaOrEntry, value, message.guild)]);
-		}
+		const entry = this.gateway.schema.get(key);
+		if (!entry) throw message.language.get('COMMAND_CONF_GET_NOEXT', key);
 
-		return message.replyLocale('COMMAND_CONF_SERVER', [
-			key ? `: ${key.split('.').map(toTitleCase).join('/')}` : '',
-			codeBlock('asciidoc', this.displayFolder(value as SettingsFolder))
-		]);
+		const value = message.author.settings.get(key);
+		return message.sendLocale('COMMAND_CONF_GET', [key, this.displayEntry(entry, value, message.guild)]);
 	}
 
 	public async set(message: Message, [key, valueToSet]: [string, string]): Promise<Message[]> {
 		try {
-			const [update] = await message.author.settings.update(key, valueToSet, { onlyConfigurable: true, arrayAction: 'add' });
-			return message.replyLocale('COMMAND_CONF_UPDATED', [key, this.displayEntry(update.entry, update.next, message.guild)]);
+			const [update] = await message.author.settings.update(key, valueToSet, { arrayAction: 'add' });
+			return message.sendLocale('COMMAND_CONF_UPDATED', [key, this.displayEntry(update.entry, update.next, message.guild)]);
 		} catch (error) {
 			throw String(error);
 		}
@@ -54,8 +52,8 @@ export default class extends Command {
 
 	public async remove(message: Message, [key, valueToRemove]: [string, string]): Promise<Message[]> {
 		try {
-			const [update] = await message.author.settings.update(key, valueToRemove, { onlyConfigurable: true, arrayAction: 'remove' });
-			return message.replyLocale('COMMAND_CONF_UPDATED', [key, this.displayEntry(update.entry, update.next, message.guild as Guild)]);
+			const [update] = await message.author.settings.update(key, valueToRemove, { arrayAction: 'remove' });
+			return message.sendLocale('COMMAND_CONF_UPDATED', [key, this.displayEntry(update.entry, update.next, message.guild)]);
 		} catch (error) {
 			throw String(error);
 		}
@@ -64,43 +62,30 @@ export default class extends Command {
 	public async reset(message: Message, [key]: [string]): Promise<Message[]> {
 		try {
 			const [update] = await message.author.settings.reset(key);
-			return message.replyLocale('COMMAND_CONF_RESET', [key, this.displayEntry(update.entry, update.next, message.guild as Guild)]);
+			return message.sendLocale('COMMAND_CONF_RESET', [key, this.displayEntry(update.entry, update.next, message.guild)]);
 		} catch (error) {
 			throw String(error);
 		}
 	}
 
-	public init(): void {
-		const { schema } = this.client.gateways.get('users') as GatewayStorage;
-		if (this.initFolderConfigurableRecursive(schema)) this.configurableSchemaKeys.set(schema.path, schema);
-	}
-
-	private displayFolder(settings: SettingsFolder): string {
+	private displayFolder(settings: Settings): string {
 		const array = [];
-		const folders = [];
 		const sections = new Map<string, string[]>();
 		let longest = 0;
-		for (const [key, value] of settings.schema.entries()) {
-			if (!this.configurableSchemaKeys.has(value.path)) continue;
+		for (const [key, value] of settings.gateway.schema.entries()) {
+			const values = sections.get(value.type) || [];
+			values.push(key);
 
-			if (value.type === 'Folder') {
-				folders.push(`// ${key}`);
-			} else {
-				const values = sections.get(value.type) || [];
-				values.push(key);
-
-				if (key.length > longest) longest = key.length;
-				if (values.length === 1) sections.set(value.type, values);
-			}
+			if (key.length > longest) longest = key.length;
+			if (values.length === 1) sections.set(value.type, values);
 		}
-		if (folders.length) array.push('= Folders =', ...folders.sort(), '');
 		if (sections.size) {
 			for (const keyType of [...sections.keys()].sort()) {
 				array.push(`= ${toTitleCase(keyType)}s =`,
 					// eslint-disable-next-line @typescript-eslint/no-non-null-assertion
 					...sections.get(keyType)!.sort().map(key =>
 						// eslint-disable-next-line @typescript-eslint/no-non-null-assertion
-						`${key.padEnd(longest)} :: ${this.displayEntry(settings.schema.get(key) as SchemaEntry, settings.get(key), settings.base!.target as Guild)}`),
+						`${key.padEnd(longest)} :: ${this.displayEntry(settings.gateway.schema.get(key) as SchemaEntry, settings.get(key), null)}`),
 					'');
 			}
 		}
@@ -122,19 +107,6 @@ export default class extends Command {
 		return values.length === 0 ?
 			'None' :
 			`[ ${values.map(value => this.displayEntrySingle(entry, value, guild)).join(' | ')} ]`;
-	}
-
-	private initFolderConfigurableRecursive(folder: Schema): boolean {
-		const previousConfigurableCount = this.configurableSchemaKeys.size;
-		for (const value of folder.values()) {
-			if (SchemaFolder.is(value)) {
-				if (this.initFolderConfigurableRecursive(value)) this.configurableSchemaKeys.set(value.path, value);
-			} else if (value.configurable) {
-				this.configurableSchemaKeys.set(value.path, value);
-			}
-		}
-
-		return previousConfigurableCount !== this.configurableSchemaKeys.size;
 	}
 
 }
