+++ src/lib/util/QueryBuilder.ts
@@ -104,7 +104,7 @@ export class QueryBuilder extends Map<string, QueryBuilderValue> {
 		const parsedDefault = this.serialize(schemaEntry.default, schemaEntry, datatype);
 		const type = typeof datatype.type === 'function' ? datatype.type(schemaEntry) : datatype.type;
 		const parsedDatatype = schemaEntry.array ? datatype.array(type) : type;
-		return datatype.formatDatatype(schemaEntry.path, parsedDatatype, parsedDefault);
+		return datatype.formatDatatype(schemaEntry.key, parsedDatatype, parsedDefault);
 	}
 
 	/**
